requirements.txt
chromadb==1.0.7
crewai==0.114.0
crewai_tools==0.42.0
langchain==0.3.24
mistralai==1.7.0
PyPDF2==3.0.1
python-dotenv==1.1.0
Requests==2.32.3
sentence_transformers==4.1.0
streamlit==1.44.1
